# openapi.yaml
openapi: 3.1.0
info:
  title: Traccar
  description: Traccar GPS tracking server API documentation. To use the API you need to have a server instance. For testing purposes you can use one of free [demo servers](https://www.traccar.org/demo-server/). For production use you can install your own server or get a [subscription service](https://www.traccar.org/product/tracking-server/).
  contact:
    name: Traccar Support
    url: https://www.traccar.org/
    email: support@traccar.org
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  version: '6.12.2'
servers:
  - url: https://demo.traccar.org/api
    description: Demo Server 1
  - url: https://demo2.traccar.org/api
    description: Demo Server 2
  - url: https://demo3.traccar.org/api
    description: Demo Server 3
  - url: https://demo4.traccar.org/api
    description: Demo Server 4
  - url: https://server.traccar.org/api
    description: Subscription Server
  - url: http://{host}:{port}/api
    description: Other Server
    variables:
      host:
        default: localhost
      port:
        enum:
          - '8082'
          - '80'
        default: '8082'
security:
  - BasicAuth: []
  - ApiKey: []
...
paths:
...
  /permissions:
    post:
      summary: Link an Object to another Object
      tags:
        - Permissions
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Permission'
        required: true
      responses:
        '204':
          description: No Content
          content: {}
        '400':
          description: No permission
          content: {}
      x-codegen-request-body-name: body
    delete:
      summary: Unlink an Object from another Object
      tags:
        - Permissions
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Permission'
        required: true
      responses:
        '204':
          description: No Content
          content: {}
      x-codegen-request-body-name: body
...
  /users:
    get:
      summary: Fetch a list of Users
      tags:
        - Users
      parameters:
        - name: userId
          in: query
          description: Can only be used by admin or manager users
          schema:
            type: string
        - name: limit
          in: query
          description: Limit the number of returned results
          schema:
            type: integer
        - name: offset
          in: query
          description: Offset for pagination
          schema:
            type: integer
        - name: keyword
          in: query
          description: Search keyword filter (searches name, email)
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        '400':
          description: No Permission
          content: {}
    post:
      summary: Create a User
      tags:
        - Users
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
      x-codegen-request-body-name: body
...
  /users/{id}:
    put:
      summary: Update a User
      tags:
        - Users
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
      x-codegen-request-body-name: body
    delete:
      summary: Delete a User
      tags:
        - Users
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: No Content
          content: {}
...
  /attributes/computed:
    get:
      summary: Fetch a list of Attributes
      description: Without params, it returns a list of Attributes the user has access to
      tags:
        - Attributes
      parameters:
        - name: all
          in: query
          description: Can only be used by admins or managers to fetch all entities
          schema:
            type: boolean
        - name: userId
          in: query
          description: Standard users can use this only with their own _userId_
          schema:
            type: integer
        - name: deviceId
          in: query
          description: >-
            Standard users can use this only with _deviceId_s, they have access
            to
          schema:
            type: integer
        - name: groupId
          in: query
          description: >-
            Standard users can use this only with _groupId_s, they have access
            to
          schema:
            type: integer
        - name: refresh
          in: query
          schema:
            type: boolean
        - name: limit
          in: query
          description: Limit the number of returned results
          schema:
            type: integer
        - name: offset
          in: query
          description: Offset for pagination
          schema:
            type: integer
        - name: keyword
          in: query
          description: Search keyword filter
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Attribute'
    post:
      summary: Create an Attribute
      tags:
        - Attributes
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Attribute'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Attribute'
      x-codegen-request-body-name: body
  /attributes/computed/{id}:
    put:
      summary: Update an Attribute
      tags:
        - Attributes
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Attribute'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Attribute'
      x-codegen-request-body-name: body
    delete:
      summary: Delete an Attribute
      tags:
        - Attributes
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: No Content
          content: {}
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Unique user identifier
        name:
          type: string
          description: User display name
        email:
          type: string
          description: Email address used for login and notifications
        phone:
          type: string
          nullable: true
          description: Contact phone number for alerts
        readonly:
          type: boolean
          description: When true, the user cannot change settings
        administrator:
          type: boolean
          description: Grants full administrative privileges when enabled
        map:
          type: string
          nullable: true
          description: Preferred default map layer for the user
        latitude:
          type: number
          description: Default map center latitude for this user
        longitude:
          type: number
          description: Default map center longitude for this user
        zoom:
          type: integer
          description: Default map zoom level on login
        password:
          type: string
          description: Password for user authentication
        coordinateFormat:
          type: string
          nullable: true
          description: Preferred coordinate display format
        disabled:
          type: boolean
          description: Indicates whether the user account is disabled
        expirationTime:
          type: string
          description: in ISO 8601 format. eg. `1963-11-22T18:30:00Z`
          format: date-time
          nullable: true
        deviceLimit:
          type: integer
          description: Maximum number of devices the user can manage
        userLimit:
          type: integer
          description: Maximum number of subordinate users
        deviceReadonly:
          type: boolean
          description: Restricts user from modifying device attributes
        limitCommands:
          type: boolean
          description: Prevents user from sending unsupported commands
        fixedEmail:
          type: boolean
          description: Locks the email field to avoid changes
        poiLayer:
          type: string
          nullable: true
          description: External POI layer configured for the user
        attributes:
          type: object
          description: Additional custom user attributes
          properties: {}
...
    Permission:
      type: object
      properties:
        userId:
          type: integer
          format: int64
          description: User id, can be only first parameter
        deviceId:
          type: integer
          format: int64
          description: >-
            Device id, can be first parameter or second only in combination with
            userId
        groupId:
          type: integer
          format: int64
          description: >-
            Group id, can be first parameter or second only in combination with
            userId
        geofenceId:
          type: integer
          format: int64
          description: Geofence id, can be second parameter only
        notificationId:
          type: integer
          format: int64
          description: Notification id, can be second parameter only
        calendarId:
          type: integer
          format: int64
          description: >-
            Calendar id, can be second parameter only and only in combination
            with userId
        attributeId:
          type: integer
          format: int64
          description: Computed attribute id, can be second parameter only
        driverId:
          type: integer
          format: int64
          description: Driver id, can be second parameter only
        managedUserId:
          type: integer
          format: int64
          description: >-
            User id, can be second parameter only and only in combination with
            userId
        commandId:
          type: integer
          format: int64
          description: Saved command id, can be second parameter only
      description: >-
        This is a permission map that contain two object indexes. It is used to
        link/unlink objects. Order is important. Example: { deviceId:8,
        geofenceId: 16 }
...
    Attribute:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Unique computed attribute identifier
        description:
          type: string
          description: Human readable name of the attribute
        attribute:
          type: string
          description: Attribute name used in expressions
        expression:
          type: string
          description: Expression that defines how the attribute is calculated
        type:
          type: string
          description: String|Number|Boolean
...
  securitySchemes:
    BasicAuth:
      type: http
      scheme: basic
    ApiKey:
      type: http
      scheme: bearer